+++ beacon-chain/deprecated-sync/initial-sync/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
         "sync_blocks.go",
         "sync_state.go",
     ],
-    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync",
+    importpath = "github.com/prysmaticlabs/prysm/beacon-chain/deprecated-sync/initial-sync",
     visibility = ["//beacon-chain:__subpackages__"],
     deps = [
         "//beacon-chain/blockchain:go_default_library",
