+++ beacon-chain/deprecated-sync/initial-sync/helpers.go
