+++ beacon-chain/deprecated-sync/initial-sync/service.go
