+++ beacon-chain/deprecated-sync/initial-sync/sync_blocks.go
