+++ beacon-chain/deprecated-sync/initial-sync/sync_state.go
