+++ beacon-chain/deprecated-sync/regular_sync.go
