+++ beacon-chain/deprecated-sync/regular_sync_test.go
