+++ beacon-chain/deprecated-sync/service.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 
 	"github.com/prysmaticlabs/prysm/beacon-chain/db"
+	initialsync "github.com/prysmaticlabs/prysm/beacon-chain/deprecated-sync/initial-sync"
 	"github.com/prysmaticlabs/prysm/beacon-chain/operations"
-	initialsync "github.com/prysmaticlabs/prysm/beacon-chain/sync/initial-sync"
 	"github.com/sirupsen/logrus"
 )
